Update preview download section
