+++ src/ensembl/src/content/app/custom-download/containers/content/preview-download/PreviewDownload.scss
@@ -10,23 +10,37 @@
     padding: 8px 0 8px 7px;
     font-size: 14px;
     vertical-align: top;
-
-    div {
-      margin-bottom: 5px;
-    }
   }
 }
 
 .previewDownloadHeader {
   border-collapse: collapse;
   min-width: 50%;
-  border-bottom: 2px solid $grey;
+  border-bottom: 1px solid $grey;
   color: $grey;
 
   td {
-    padding: 8px 0 8px 8px;
+    padding: 8px 0 0px 8px;
     white-space: nowrap;
     text-overflow: ellipsis;
+    position: relative;
+  }
+
+  td.species {
+    width: 20%;
+    color: $blue;
+    line-height: 35px;
+  }
+
+  td.example {
+    width: 40%;
+    padding-left: 0;
+  }
+
+  td.filters {
+    width: 40%;
+    color: $blue;
+    line-height: 35px;
   }
 }
 
@@ -36,16 +50,25 @@
   position: relative;
 }
 
-.changeLink {
-  color: $blue;
-  cursor: pointer;
-  margin-top: 20px;
-}
-
 .closeButton {
   height: 15px;
   width: 15px;
   position: absolute;
   right: 20px;
   top: 20px;
+}
+
+.exampleDataPanel {
+  width: calc(100% - 30px);
+}
+
+.resultCounter {
+  position: absolute;
+  top: 10px;
+  right: 100px;
+  color: $black;
+  font-size: 13px;
+  span {
+    font-weight: 700;
+  }
 }
