--- src/ensembl/src/content/app/custom-download/containers/content/result-loader/ResultLoader.scss
@@ -1,61 +1,0 @@
-@import 'src/styles/common';
-
-.wrapper {
-  padding-right: 16px;
-  padding-left: 50px;
-  margin-top: -15px;
-  height: calc(100vh - 120px);
-  overflow-y: auto;
-  margin-right: 4px;
-}
-
-.resultCard {
-  min-height: 75px;
-  border: 1px solid $grey;
-  padding: 15px 5px;
-  background: $white;
-  margin-bottom: 5px;
-  position: relative;
-  display: block;
-
-  .resultLine {
-    width: 100%;
-    position: relative;
-    clear: both;
-    .lineHeader {
-      text-align: right;
-      color: $dark-grey;
-      width: 25%;
-      max-width: 200px;
-      text-overflow: ellipsis;
-      overflow: hidden;
-      float: left;
-    }
-
-    .lineValue {
-      color: $black;
-      width: 75%;
-      display: inline-block;
-      padding-left: 5px;
-    }
-  }
-}
-
-.loaderWrapper {
-  text-align: center;
-  padding-right: 16px;
-  padding-left: 50px;
-  margin-top: -15px;
-  max-height: calc(100vh - 210px);
-  overflow-y: auto;
-  margin-right: 4px;
-
-  div {
-    top: 40%;
-    position: absolute;
-  }
-}
-
-.noResult {
-  margin-top: 45px;
-}
